Moved Permission and PermissionEnum to project_types
