+++ src/project_types/mod.rs
@@ -1,9 +1,11 @@
 
+mod permissions;
 mod project;
 mod sequence;
 mod sequence_section;
 mod user;
 
+pub use self::permissions::{Permission, PermissionEnum};
 pub use self::project::Project;
 pub use self::sequence::Sequence;
 pub use self::sequence_section::SequenceSection;
